Eliminado registerListItem de list components e adicionado em components de dashboard de admin
